need to make things go in order
