+++ index.js
@@ -1,10 +1,10 @@
 'use strict';
 
-const fileReader = require('./lib/reader.js');
+const fileReader = require('./lib/print-files.js');
 
-let textFiles = [__dirname + '/data/data.txt', __dirname + '/data/data2.txt', __dirname + '/data/data3.txt'];
+// let textFiles = [__dirname + '/data/data.txt', __dirname + '/data/data2.txt', __dirname + '/data/data3.txt'];
 
-fileReader(textFiles, (err, data) => {
+fileReader((err, data) => {
   if(err) console.log(err);
   console.log(data);
 });
